@@ -1,6 +1,7 @@
 ﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved. 
 // SPDX - License - Identifier: Apache - 2.0
-// snippet-start:[dynamodb.dotnet35.GetLowProductStockGSI]
+
+// snippet-start:[dynamodb.dotnet35.GetLowProductStockGSIExample]
 using System;
 using System.Collections.Generic;
 using System.Configuration;
@@ -44,9 +45,9 @@ namespace GetLowProductStockGSI
             var region = "";
             var table = "";
             var index = "";
-            string minimum = "";            
+            string minimum = "";
 
-            // Get default values from config file
+            // Get default values from config file.
             var efm = new ExeConfigurationFileMap
             {
                 ExeConfigFilename = configfile
@@ -113,4 +114,4 @@ namespace GetLowProductStockGSI
         }
     }
 }
-// snippet-end:[dynamodb.dotnet35.GetLowProductStockGSI]
+// snippet-end:[dynamodb.dotnet35.GetLowProductStockGSIExample]
