@@ -3,19 +3,19 @@
 
 namespace HighLevelMappingArbitraryDataExample
 {
-  // snippet-start:[dynamodb.dotnet35.HighLevelMappingArbitraryData.DimensionType]
+    // snippet-start:[dynamodb.dotnet35.HighLevelMappingArbitraryData.DimensionType]
 
-  /// <summary>
-  /// Defines the dimensions of a book.
-  /// </summary>
-  public class DimensionType
-  {
-    public decimal Length { get; set; }
+    /// <summary>
+    /// Defines the dimensions of a book.
+    /// </summary>
+    public class DimensionType
+    {
+        public decimal Length { get; set; }
 
-    public decimal Height { get; set; }
+        public decimal Height { get; set; }
 
-    public decimal Thickness { get; set; }
-  }
+        public decimal Thickness { get; set; }
+    }
 
-  // snippet-end:[dynamodb.dotnet35.HighLevelMappingArbitraryData.DimensionType]
+    // snippet-end:[dynamodb.dotnet35.HighLevelMappingArbitraryData.DimensionType]
 }
