@@ -3,34 +3,34 @@
 
 namespace HighLevelQueryAndScanExample
 {
-  using System;
-  using Amazon.DynamoDBv2.DataModel;
+    using System;
+    using Amazon.DynamoDBv2.DataModel;
 
-  // snippet-start:[dynamodb.dotnet35.HighLevelQueryAndScan.Reply]
+    // snippet-start:[dynamodb.dotnet35.HighLevelQueryAndScan.Reply]
 
-  /// <summary>
-  /// Maps information to the Amazon DynamoDB table Reply.
-  /// </summary>
-  [DynamoDBTable("Reply")]
-  public class Reply
-  {
-    [DynamoDBHashKey] // Partition key
-    public string Id { get; set; }
+    /// <summary>
+    /// Maps information to the Amazon DynamoDB table Reply.
+    /// </summary>
+    [DynamoDBTable("Reply")]
+    public class Reply
+    {
+        [DynamoDBHashKey] // Partition key
+        public string Id { get; set; }
 
-    [DynamoDBRangeKey] // Sort key
-    public DateTime ReplyDateTime { get; set; }
+        [DynamoDBRangeKey] // Sort key
+        public DateTime ReplyDateTime { get; set; }
 
-    // Properties included implicitly.
-    public string Message { get; set; }
+        // Properties included implicitly.
+        public string Message { get; set; }
 
-    // Explicit property mapping with object persistence model attributes.
-    [DynamoDBProperty("LastPostedBy")]
-    public string PostedBy { get; set; }
+        // Explicit property mapping with object persistence model attributes.
+        [DynamoDBProperty("LastPostedBy")]
+        public string PostedBy { get; set; }
 
-    // Property to store version number for optimistic locking.
-    [DynamoDBVersion]
-    public int? Version { get; set; }
-  }
+        // Property to store version number for optimistic locking.
+        [DynamoDBVersion]
+        public int? Version { get; set; }
+    }
 
-  // snippet-end:[dynamodb.dotnet35.HighLevelQueryAndScan.Reply]
+    // snippet-end:[dynamodb.dotnet35.HighLevelQueryAndScan.Reply]
 }
