@@ -1,6 +1,7 @@
 ﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX - License - Identifier: Apache - 2.0
-// snippet-start:[dynamodb.dotnet35.LowLevelBatchWrite]
+
+// snippet-start:[dynamodb.dotnet35.LowLevelBatchWriteExample]
 using System;
 using System.Collections.Generic;
 using Amazon.DynamoDBv2;
@@ -133,4 +134,4 @@ namespace LowLevelBatchWrite
         }
     }
 }
-// snippet-end:[dynamodb.dotnet35.LowLevelBatchWrite]
+// snippet-end:[dynamodb.dotnet35.LowLevelBatchWriteExample]
