@@ -1,6 +1,7 @@
 ﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX - License - Identifier: Apache - 2.0
-// snippet-start:[dynamodb.dotnet35.LowLevelScan]
+
+// snippet-start:[dynamodb.dotnet35.LowLevelScanExample]
 using System;
 using System.Collections.Generic;
 using Amazon.DynamoDBv2;
@@ -72,4 +73,4 @@ namespace LowLevelScan
         }
     }
 }
-// snippet-end:[dynamodb.dotnet35.LowLevelScan]
+// snippet-end:[dynamodb.dotnet35.LowLevelScanExample]
